filesystem is not resized when restoring from snapshot/cloning to larger size than origin (#100)
* remove unnecessary code

VolumeMountUtils is already instantiated in init() of ibmcsidriver
module so there is no need for this line of code.

* allow creating driver with fake commands

In order to allow testing of code that includes command sequence we
need to change the driver so it can accept a list of fake commands.

* resize fs in NodeStageVolume

This is the actual fix for resizing filesystem when volume is restored
from a snapshot to larger size

* resize with mounter during node expand

* cleanup resize code

* update to latest ibm-csi-common